# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 工业设备智能运维系统 API 契约
  version: 2.0.0
  description: |
    本文件是成员 A（设备资产与监测）、B（健康评估与预警）、
    C（维修工单与备件）、D（身份、通知与集成）共同遵守的唯一 HTTP 契约。

    面向前端的请求统一通过本地网关 http://localhost:8080。
    服务间调用可使用各模块直连地址，并必须携带 X-Trace-Id、
    Idempotency-Key（写操作）和 X-Internal-Token。

    字段、枚举、状态码和路径的任何变更必须先修改本文件，
    再由提供方和调用方共同评审后实现。
servers:
  - url: http://localhost:8080
    description: 本地统一网关（成员 D 维护路由）
...
security:
  - bearerAuth: []
paths:
  /api/v1/equipment:
    get:
      tags: [A-设备监测]
      summary: 分页查询设备
      operationId: listEquipment
      x-contract-id: A-API-01
      x-owner-member: MEMBER_A
      x-owner-port: 8101
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - name: keyword
          in: query
          description: 按设备编号或名称模糊查询
          schema:
            type: string
            maxLength: 50
        - name: productionLineId
          in: query
          schema:
            type: string
            maxLength: 32
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EquipmentStatus'
        - $ref: '#/components/parameters/PageParameter'
        - $ref: '#/components/parameters/PageSizeParameter'
      responses:
        '200':
          description: 查询成功；items 为空表示没有匹配数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [A-设备监测]
      summary: 新建设备档案
      operationId: createEquipment
      x-contract-id: A-API-02
      x-owner-member: MEMBER_A
      x-owner-port: 8101
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEquipmentRequest'
      responses:
        '201':
          description: 创建成功；重复幂等请求返回同一设备
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/equipment/{equipmentId}:
    parameters:
      - $ref: '#/components/parameters/EquipmentIdParameter'
    get:
      tags: [A-设备监测]
      summary: 查询单台设备完整档案
      operationId: getEquipmentById
      x-contract-id: C-INT-01
      x-owner-member: MEMBER_A
      x-caller-members: [MEMBER_B, MEMBER_C]
      x-owner-port: 8101
      security:
        - bearerAuth: []
        - internalToken: []
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [A-设备监测]
      summary: 修改设备档案
      operationId: updateEquipment
      x-contract-id: A-API-03
      x-owner-member: MEMBER_A
      x-owner-port: 8101
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEquipmentRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/warnings:
    get:
      tags: [B-故障预警]
      summary: 分页查询故障预警
      operationId: listWarnings
      x-contract-id: B-API-01
      x-owner-member: MEMBER_B
      x-owner-port: 8102
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - name: equipmentId
          in: query
          schema:
            $ref: '#/components/schemas/EquipmentId'
        - name: riskLevel
          in: query
          schema:
            $ref: '#/components/schemas/RiskLevel'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WarningStatus'
        - $ref: '#/components/parameters/PageParameter'
        - $ref: '#/components/parameters/PageSizeParameter'
      responses:
        '200':
          description: 按 warningAt 降序返回
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarningPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/warnings/{warningId}:
    parameters:
      - $ref: '#/components/parameters/WarningIdParameter'
    get:
      tags: [B-故障预警]
      summary: 查询预警详情
      operationId: getWarningById
      x-contract-id: B-API-02
      x-owner-member: MEMBER_B
      x-owner-port: 8102
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warning'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/warnings/{warningId}/acknowledgements:
    parameters:
      - $ref: '#/components/parameters/WarningIdParameter'
    post:
      tags: [B-故障预警]
      summary: 确认预警
      operationId: acknowledgeWarning
      x-contract-id: B-API-03
      x-owner-member: MEMBER_B
      x-owner-port: 8102
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarningAcknowledgementRequest'
      responses:
        '200':
          description: 确认成功；返回更新后的预警
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warning'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录接口签发的访问令牌
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: 本地开发由 INTERNAL_API_TOKEN 环境变量提供
  parameters:
    TraceIdHeader:
      name: X-Trace-Id
      in: header
      required: false
      description: 调用方可传；缺失时网关生成并回传
      schema:
        type: string
        minLength: 8
        maxLength: 64
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID；同一业务写请求重试时保持不变
      schema:
        type: string
        format: uuid
    PageParameter:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParameter:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    EquipmentIdParameter:
      name: equipmentId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/EquipmentId'
    WarningIdParameter:
      name: warningId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/WarningId'
...
  responses:
...
    BadRequest:
      description: 字段缺失、格式错误、枚举错误或查询参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未登录、JWT 无效或内部令牌无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 身份有效但缺少该操作所需权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 指定业务对象不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 当前状态不允许操作、版本冲突或幂等键对应不同请求
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务内部错误；不得向客户端暴露堆栈或数据库细节
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    EquipmentId:
      type: string
      pattern: '^EQ-[0-9]{6}$'
      example: EQ-000001
    WarningId:
      type: string
      pattern: '^WARN-[0-9]{8}-[0-9]{4}$'
      example: WARN-20260917-0001
    OrderId:
      type: string
      pattern: '^WO-[0-9]{8}-[0-9]{4}$'
      example: WO-20260917-0001
...
    UserId:
      type: string
      pattern: '^USER-[A-Z0-9-]{1,27}$'
      example: USER-C-002
    EquipmentStatus:
      type: string
      enum: [RUNNING, WARNING, STOPPED, MAINTAINING, TRIAL_RUNNING]
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    WarningStatus:
      type: string
      enum: [OPEN, ACKNOWLEDGED, LINKED_TO_ORDER, RESOLVED, FALSE_POSITIVE]
...
    Equipment:
      type: object
      required:
        - equipmentId
        - name
        - equipmentType
        - productionLineId
        - location
        - currentStatus
        - enabled
        - version
        - createdAt
        - updatedAt
      properties:
        equipmentId:
          $ref: '#/components/schemas/EquipmentId'
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: 一号数控机床
        equipmentType:
          type: string
          minLength: 1
          maxLength: 50
          example: CNC
        productionLineId:
          type: string
          minLength: 1
          maxLength: 32
          example: LINE-01
        location:
          type: string
          minLength: 1
          maxLength: 100
          example: 一车间 A 区
        manufacturer:
          type: [string, 'null']
          maxLength: 100
          example: 示例机床厂
        model:
          type: [string, 'null']
          maxLength: 100
          example: CNC-X1000
        responsibleDepartment:
          type: [string, 'null']
          maxLength: 100
          example: 机加车间
        currentStatus:
          $ref: '#/components/schemas/EquipmentStatus'
        enabled:
          type: boolean
          example: true
        version:
          type: integer
          minimum: 0
          example: 3
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    CreateEquipmentRequest:
      type: object
      required: [name, equipmentType, productionLineId, location]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        equipmentType:
          type: string
          minLength: 1
          maxLength: 50
        productionLineId:
          type: string
          minLength: 1
          maxLength: 32
        location:
          type: string
          minLength: 1
          maxLength: 100
        manufacturer:
          type: [string, 'null']
          maxLength: 100
        model:
          type: [string, 'null']
          maxLength: 100
        responsibleDepartment:
          type: [string, 'null']
          maxLength: 100
      additionalProperties: false
    UpdateEquipmentRequest:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion:
          type: integer
          minimum: 0
        name:
          type: string
          minLength: 1
          maxLength: 100
        productionLineId:
          type: string
          minLength: 1
          maxLength: 32
        location:
          type: string
          minLength: 1
          maxLength: 100
        responsibleDepartment:
          type: [string, 'null']
          maxLength: 100
        enabled:
          type: boolean
      minProperties: 2
      additionalProperties: false
    EquipmentPage:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Equipment'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
...
    Warning:
      type: object
      required:
        - warningId
        - equipmentId
        - riskLevel
        - healthScore
        - suspectedFault
        - recommendedAction
        - status
        - warningAt
        - modelVersion
        - linkedOrderId
        - acknowledgedBy
        - acknowledgedAt
        - version
      properties:
        warningId:
          $ref: '#/components/schemas/WarningId'
        equipmentId:
          $ref: '#/components/schemas/EquipmentId'
        riskLevel:
          $ref: '#/components/schemas/RiskLevel'
        healthScore:
          type: number
          minimum: 0
          maximum: 100
        suspectedFault:
          type: string
          maxLength: 200
        recommendedAction:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/WarningStatus'
        warningAt:
          type: string
          format: date-time
        modelVersion:
          type: string
          maxLength: 32
        linkedOrderId:
          oneOf:
            - $ref: '#/components/schemas/OrderId'
            - type: 'null'
        acknowledgedBy:
          oneOf:
            - $ref: '#/components/schemas/UserId'
            - type: 'null'
        acknowledgedAt:
          type: [string, 'null']
          format: date-time
        version:
          type: integer
          minimum: 0
      additionalProperties: false
    WarningPage:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Warning'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
    WarningAcknowledgementRequest:
      type: object
      required: [operatorId, expectedVersion]
      properties:
        operatorId:
          $ref: '#/components/schemas/UserId'
        expectedVersion:
          type: integer
          minimum: 0
        comment:
          type: [string, 'null']
          maxLength: 300
      additionalProperties: false
...
    FieldViolation:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
      additionalProperties: false
    ErrorResponse:
      type: object
      required: [code, message, traceId, timestamp, details]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          example: EQUIPMENT_NOT_FOUND
        message:
          type: string
          example: 设备不存在
        traceId:
          type: string
          minLength: 8
          maxLength: 64
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
      additionalProperties: false